Fixes #347: Add support for media import syntax
Users can do things like `import imageURL from './image.svg';` when
using react-scripts because it configures a webpack plugin to do so.

This ends up being a URL in the bundled JS code that starts with
'/static/media/' prefix.

Because we are deploying the solution as a portlet we need to namespace
those URLs, so we substitute the prefix in all bundled JS files. This is
far from a perfect approach, because we substitute each occurrence,
which can lead to false positives while, on the other hand, we are not
capable of substituting the prefix if it is split or dinamically
composed.

Unfortunately, there's no way to do it better because the most semantic
action we can take is to parse the AST tree, but still we would be at
syntactic level so we don't know which ones to substitute and which ones
to leave untouched :-(.

We could do this better if we were able to tune the react-scripts build
because that way we would know for sure which URLs we must substitute.
But given that it is impossible because react-scripts cannot be
extended, we are left with just this imperfect possibility.